BXMSPROD-952 workflow-informaton-reader created
